Changed behavior of `Services.start`
It did reject if a service failed to start, but now it does not. Instead
it resolves with a list of services that successfully started and those
that did not.

If you need a bunch of services to have been guaranteed to successfully
start, then you should wrap that code in `Services.ready`, which will
not resolve unless all the required services are successfully started.